fix: improve a few things - split code - drop some rubbish function - add useMemo - rename action to Capital Letter - drop useless dep arr
