+++ resources/assets/js/components/atom/PreviewInput/index.js
@@ -11,7 +11,7 @@ function PreviewInput({ keyName, placeHolder, formType }) {
     const targetForm = form[formType];
     const { keysTable } = DataMapping.transObjToKeysTable(
         targetForm,
-        DataMapping.action.getInitStudy
+        DataMapping.action.GetInitStudy
     );
     const obj = result(targetForm, keysTable[keyName][0], {}); //already handle exception
     const data = !obj.value
