+++ resources/assets/js/model/reducer/post/function/index.js
@@ -1,0 +1,5 @@
+import { searchingAlgorithm } from './searchingAlgo';
+
+export default {
+    searchingAlgorithm: searchingAlgorithm,
+};
